+++ src/swfexport.cpp
@@ -89,8 +89,8 @@ void MolDisplayWin::CreateFrameMovie(wxString &filePath,
 	SWFDisplayItem *di = NULL;
 	SWFDisplayItem *textDI = NULL;
 	SWFDisplayItem *di2 = NULL;
-	long width;
-	long height;
+	int width;
+	int height;
 	int AnimateTime = 10*Prefs->GetAnimateTime();
 	bool killEPlotWin = false;
 	int savedEPlotWidth, savedEPlotHeight;
@@ -253,8 +253,8 @@ void MolDisplayWin::CreateModeMovie(wxString &filePath) {
 	SWFBitmap *bm = NULL;
 	SWFDisplayItem *di = NULL;
 	SWFDisplayItem *textDI = NULL;
-	long width;
-	long height;
+	int width;
+	int height;
 	int AnimateTime = 10*Prefs->GetAnimateTime();
 	wxStandardPathsBase & gStdPaths = wxStandardPaths::Get();
 #if wxCHECK_VERSION(2, 8, 0)
